+++ browser/pages/confirm/controller.js
@@ -23,13 +23,7 @@ class ConfirmController {
     $scope.platform = Platform.OS;
     $scope.detectionStyle = false;
     $scope.virtualization = true;
-
-    for (let [key, value] of this.installerDataSvc.allInstallables().entries()) {
-      $scope.checkboxModel[key] = value;
-      $scope.$watch(`checkboxModel.${key}.selectedOption`, function watchInstallSelectionChange() {
-        $scope.checkboxModel[key].validateVersion();
-      });
-    }
+    $scope.downloadComp = this.detectDownloadedComponents();
 
     const selectAllLabel = 'Select All Components';
 
@@ -61,31 +55,9 @@ class ConfirmController {
 
     $scope.isConfigurationValid = this.isConfigurationValid;
 
-    $scope.$watch('$viewContentLoaded', this.initPage.bind(this));
+    // $scope.$watch('$viewContentLoaded', this.initPage.bind(this));
 
-    this.electron.remote.getCurrentWindow().addListener('focus', ()=>this.timeout(this.activatePage.bind(this), true));
-
-    $scope.updateTotalDownloadSize = () => {
-      let checkboxModel = this.sc.checkboxModel;
-      let finalSize = 0;
-      for (let key in checkboxModel) {
-        if (checkboxModel[key].size && checkboxModel[key].selectedOption == 'install') {
-          finalSize += checkboxModel[key].size;
-        }
-      }
-      return finalSize;
-    };
-
-    $scope.updateTotalDiskSpace = () => {
-      let checkboxModel = this.sc.checkboxModel;
-      let totalDownloadSize = $scope.updateTotalDownloadSize();
-      for (let key in checkboxModel) {
-        if (checkboxModel[key].installSize && checkboxModel[key].selectedOption == 'install') {
-          totalDownloadSize += checkboxModel[key].installSize;
-        }
-      }
-      return totalDownloadSize;
-    };
+    // this.electron.remote.getCurrentWindow().addListener('focus', ()=>this.timeout(this.activatePage.bind(this), true));
   }
 
   selectAll() {
@@ -105,102 +77,22 @@ class ConfirmController {
     }
   }
 
-  initPage() {
-    return this.detectInstalledComponents().then(()=> {
-      this.graph = ComponentLoader.loadGraph(this.installerDataSvc);
-      this.installWatchers();
-      return Promise.resolve();
-    }).then(
-      ()=> this.setIsDisabled()
-    ).catch((error)=> {
-      this.setIsDisabled();
-    });
-  }
-
-  activatePage() {
-    return this.detectInstalledComponents().then(
-      ()=> this.setIsDisabled()
-    ).catch((error)=> {
-      Logger.error(error);
-      this.setIsDisabled();
-    });
-  }
-
-  installWatchers() {
-    let graph = this.graph;
-    let nodes = graph.overallOrder() ;
-    let checkboxModel = this.sc.checkboxModel;
-    for (let node of nodes) {
-      checkboxModel[node].dependenciesOf = [];
-      for(let dependant of this.graph.dependantsOf(node)) {
-        checkboxModel[node].dependenciesOf.push(checkboxModel[dependant]);
+  detectDownloadedComponents() {
+    let downloadedComponents = [];
+    for (let value of this.installerDataSvc.allInstallables().values()) {
+      if(value.isDownloaded() && value.selectedOption == 'install') {
+        downloadedComponents.push(value);
       }
-      checkboxModel[node].references=0;
+      this.downloadComp = downloadedComponents;      
     }
-    for (let node of nodes) {
-      function watchComponent(newv, oldv) {
-        let installer = checkboxModel[node];
-        if(installer.isSelected()) {
-          for(let dep of graph.dependenciesOf(node)) {
-            let depInstaller = checkboxModel[dep];
-            if(depInstaller.isInstallable && depInstaller.references === 0 && depInstaller.isNotDetected()) {
-              depInstaller.selectedOption = 'install';
-            }
-            depInstaller.references++;
-          }
-        } else if(!installer.isSelected() && oldv === 'install') {
-          for(let dep of graph.dependenciesOf(node)) {
-            let depInstaller = checkboxModel[dep];
-            depInstaller.references--;
-            if(depInstaller.references === 0) {
-              depInstaller.selectedOption = 'detected';
-            }
-          }
-        }
-      }
-      this.sc.$watch(`checkboxModel.${node}.selectedOption`, watchComponent);
-    }
-  }
-
-  detectInstalledComponents() {
-    if(!this.isDisabled) {
-      this.isDisabled = true;
-      this.installedSearchNote = ' The system is checking if you have any installed components.';
-      let detectors = [];
-      for (var installer of this.installerDataSvc.allInstallables().values()) {
-        detectors.push(installer.detectExistingInstall());
-      }
-      this.detection = Promise.all(detectors);
-    }
-    return this.detection;
+    return this.downloadComp;
   }
 
   download(url) {
     this.electron.shell.openExternal(url);
   }
 
-  // Prep the install location path for each product, then go to the next page.
-  install() {
-    let checkboxModel = this.sc.checkboxModel;
-    if (checkboxModel.hyperv && checkboxModel.hyperv.isConfigured()) {
-      this.loader.removeComponent('virtualbox');
-    } else {
-      this.loader.removeComponent('hyperv');
-    }
-
-    let possibleComponents = ['virtualbox', 'jdk', 'devstudio', 'jbosseap', 'cygwin', 'cdk', 'kompose', 'fuseplatform', 'fuseplatformkaraf'];
-    for (let i = 0; i < possibleComponents.length; i++) {
-      let component = this.installerDataSvc.getInstallable(possibleComponents[i]);
-      if (component) {
-        possibleComponents[i] = component.getLocation();
-      } else {
-        possibleComponents[i] = undefined;
-      }
-    }
-
-    this.electron.remote.getCurrentWindow().removeAllListeners('focus');
-    this.installerDataSvc.setup(...possibleComponents);
-
+  next() {
     this.router.go(this.getNextPage());
   }
 
@@ -234,56 +126,6 @@ class ConfirmController {
     }
   }
 
-  setIsDisabled() {
-    // Uncomment the timeout to see the initial disabled view.
-    this.timeout(() => {
-      // Switch this boolean flag when the app is done looking for existing installations.
-      this.isDisabled = false;
-      this.numberOfExistingInstallations = 0;
-      // Count the number of existing installations.
-      for (var [, value] of this.installerDataSvc.allInstallables()) {
-        if (value.hasOption('detected')) {
-          ++this.numberOfExistingInstallations;
-        }
-      }
-
-      if (this.numberOfExistingInstallations == 1) {
-        this.installedSearchNote = `  We found ${this.numberOfExistingInstallations} installed component.`;
-      } else if (this.numberOfExistingInstallations > 1) {
-        this.installedSearchNote = `  We found ${this.numberOfExistingInstallations} installed components.`;
-      } else {
-        this.installedSearchNote = '';
-      }
-
-      if (this.numberOfExistingInstallations == 0) {
-        this.installInstructions = 'Select components to install.';
-      } else if (this.numberOfExistingInstallations > 0) {
-        this.installInstructions = 'Select any additional components to install.';
-      }
-    }, true);
-  }
-
-  isConfigurationValid() {
-    let result = true;
-    for (let [, value] of this.installerDataSvc.allInstallables().entries()) {
-      if(! (result = value.isConfigurationValid())) {
-        break;
-      }
-    }
-    return result && this.isAtLeastOneSelected();
-  }
-
-  isAtLeastOneSelected() {
-    let selected = false;
-    for (var [, value] of this.installerDataSvc.allInstallables()) {
-      selected = !value.isSkipped();
-      if(selected) {
-        break;
-      }
-    }
-    return selected;
-  }
-
   exit() {
     Logger.info('Closing the installer window');
     this.electron.remote.getCurrentWindow().close();
@@ -291,8 +133,8 @@ class ConfirmController {
 
   back() {
     Logger.info('Going back a page');
-    this.electron.remote.getCurrentWindow().removeAllListeners('focus');
-    this.router.go('location');
+    // this.electron.remote.getCurrentWindow().removeAllListeners('focus');
+    this.router.go('selection');
   }
 }
