+++ browser/pages/selection/controller.js
@@ -1,0 +1,305 @@
+'use strict';
+
+import Logger from '../../services/logger';
+import Platform from '../../services/platform';
+import ComponentLoader from '../../services/componentLoader';
+
+class SelectionController {
+
+  constructor($scope, $state, $timeout, installerDataSvc, electron) {
+    this.router = $state;
+    this.sc = $scope;
+    this.timeout = $timeout;
+    this.installerDataSvc = installerDataSvc;
+    this.electron = electron;
+    this.loader = new ComponentLoader(installerDataSvc);
+    this.loader.loadComponents();
+    this.installedSearchNote = '';
+    this.isDisabled = false;
+    this.numberOfExistingInstallations = 0;
+
+    this.installables = {};
+    $scope.checkboxModel = {};
+    $scope.platform = Platform.OS;
+    $scope.detectionStyle = false;
+    $scope.virtualization = true;
+
+    for (let [key, value] of this.installerDataSvc.allInstallables().entries()) {
+      $scope.checkboxModel[key] = value;
+      $scope.$watch(`checkboxModel.${key}.selectedOption`, function watchInstallSelectionChange() {
+        $scope.checkboxModel[key].validateVersion();
+      });
+    }
+
+    const selectAllLabel = 'Select All Components';
+
+    menu.insert(0, new MenuItem({
+      label: selectAllLabel,
+      click: ()=> {
+        this.sc.$apply(this.selectAll.bind(this));
+      }
+    }));
+
+    const deselectAllLabel = 'Deselect All Components';
+
+
+    menu.insert(1, new MenuItem({
+      label: deselectAllLabel,
+      click: ()=> {
+        this.sc.$apply(this.deselectAll.bind(this));
+      }
+    }));
+
+    menu.insert(2, new MenuItem({
+      label: deselectAllLabel,
+      type: 'separator'
+    }));
+
+    $scope.$on('$destroy', ()=>{
+      restoreMenu();
+    });
+
+    $scope.isConfigurationValid = this.isConfigurationValid;
+
+    $scope.$watch('$viewContentLoaded', this.initPage.bind(this));
+
+    this.electron.remote.getCurrentWindow().addListener('focus', this.activatePage.bind(this));
+
+    $scope.updateTotalDownloadSize = () => {
+      let checkboxModel = this.sc.checkboxModel;
+      let finalSize = 0;
+      for (let key in checkboxModel) {
+        if (checkboxModel[key].size && checkboxModel[key].selectedOption == 'install') {
+          finalSize += checkboxModel[key].size;
+        }
+      }
+      return finalSize;
+    };
+
+    $scope.updateTotalDiskSpace = () => {
+      let checkboxModel = this.sc.checkboxModel;
+      let totalDownloadSize = $scope.updateTotalDownloadSize();
+      for (let key in checkboxModel) {
+        if (checkboxModel[key].installSize && checkboxModel[key].selectedOption == 'install') {
+          totalDownloadSize += checkboxModel[key].installSize;
+        }
+      }
+      return totalDownloadSize;
+    };
+  }
+
+  selectAll() {
+    let checkboxModel = this.sc.checkboxModel;
+    for (let key in checkboxModel) {
+      let node = checkboxModel[key];
+      if (node.isInstallable && node.isNotDetected()) {
+        node.selectedOption = 'install';
+      }
+    }
+  }
+
+  deselectAll() {
+    let checkboxModel = this.sc.checkboxModel;
+    for (let key in checkboxModel) {
+      checkboxModel[key].selectedOption = 'detected';
+    }
+  }
+
+  initPage() {
+    return this.detectInstalledComponents().then(()=> {
+      this.graph = ComponentLoader.loadGraph(this.installerDataSvc);
+      this.installWatchers();
+      return Promise.resolve();
+    }).then(
+      ()=> this.setIsDisabled()
+    ).catch((error)=> {
+      this.setIsDisabled();
+    });
+  }
+
+  activatePage() {
+    return this.detectInstalledComponents().then(
+      ()=> this.setIsDisabled()
+    ).catch((error)=> {
+      Logger.error(error);
+      this.setIsDisabled();
+    });
+  }
+
+  installWatchers() {
+    let graph = this.graph;
+    let nodes = graph.overallOrder() ;
+    let checkboxModel = this.sc.checkboxModel;
+    for (let node of nodes) {
+      checkboxModel[node].dependenciesOf = [];
+      for(let dependant of this.graph.dependantsOf(node)) {
+        checkboxModel[node].dependenciesOf.push(checkboxModel[dependant]);
+      }
+      checkboxModel[node].references=0;
+    }
+    for (let node of nodes) {
+      function watchComponent(newv, oldv) {
+        let installer = checkboxModel[node];
+        if(installer.isSelected()) {
+          for(let dep of graph.dependenciesOf(node)) {
+            let depInstaller = checkboxModel[dep];
+            if(depInstaller.isInstallable && depInstaller.references === 0 && depInstaller.isNotDetected()) {
+              depInstaller.selectedOption = 'install';
+            }
+            depInstaller.references++;
+          }
+        } else if(!installer.isSelected() && oldv === 'install') {
+          for(let dep of graph.dependenciesOf(node)) {
+            let depInstaller = checkboxModel[dep];
+            depInstaller.references--;
+            if(depInstaller.references === 0) {
+              depInstaller.selectedOption = 'detected';
+            }
+          }
+        }
+      }
+      this.sc.$watch(`checkboxModel.${node}.selectedOption`, watchComponent);
+    }
+  }
+
+  detectInstalledComponents() {
+    if(!this.isDisabled) {
+      this.isDisabled = true;
+      this.installedSearchNote = ' The system is checking if you have any installed components.';
+      let detectors = [];
+      for (var installer of this.installerDataSvc.allInstallables().values()) {
+        detectors.push(installer.detectExistingInstall());
+      }
+      this.detection = Promise.all(detectors);
+    }
+    return this.detection;
+  }
+
+  download(url) {
+    this.electron.shell.openExternal(url);
+  }
+
+  // Prep the install location path for each product, then go to the next page.
+  install() {
+    let checkboxModel = this.sc.checkboxModel;
+    if (checkboxModel.hyperv && checkboxModel.hyperv.isConfigured()) {
+      this.loader.removeComponent('virtualbox');
+    } else {
+      this.loader.removeComponent('hyperv');
+    }
+
+    let possibleComponents = ['virtualbox', 'jdk', 'devstudio', 'jbosseap', 'cygwin', 'cdk', 'kompose', 'fuseplatform', 'fuseplatformkaraf'];
+    for (let i = 0; i < possibleComponents.length; i++) {
+      let component = this.installerDataSvc.getInstallable(possibleComponents[i]);
+      if (component) {
+        possibleComponents[i] = component.getLocation();
+      } else {
+        possibleComponents[i] = undefined;
+      }
+    }
+
+    this.electron.remote.getCurrentWindow().removeAllListeners('focus');
+    this.installerDataSvc.setup(...possibleComponents);
+
+    this.router.go(this.next());
+  }
+
+  isAccountRequired() {
+    let checkboxModel = this.sc.checkboxModel;
+    let required = false;
+    for (const key in checkboxModel) {
+      required = checkboxModel.hasOwnProperty(key)
+        && checkboxModel[key].authRequired
+        && checkboxModel[key].selectedOption == 'install';
+      if(required) {
+        break;
+      }
+    }
+    return required;
+  }
+
+  // getNextPage () {
+  //   if(this.isAccountRequired()) {
+  //     return 'confirm';
+  //   } else {
+  //     return 'confirm';
+  //   }
+  // }
+
+  // getNextButtonName () {
+  //   if(this.isAccountRequired()) {
+  //     return 'Next';
+  //   } else {
+  //     return 'Download & Install';
+  //   }
+  // }
+
+  next() {
+    this.router.go('confirm');
+  }
+
+  setIsDisabled() {
+    // Uncomment the timeout to see the initial disabled view.
+    this.timeout(() => {
+      // Switch this boolean flag when the app is done looking for existing installations.
+      this.isDisabled = false;
+      this.numberOfExistingInstallations = 0;
+      // Count the number of existing installations.
+      for (var [, value] of this.installerDataSvc.allInstallables()) {
+        if (value.hasOption('detected')) {
+          ++this.numberOfExistingInstallations;
+        }
+      }
+
+      if (this.numberOfExistingInstallations == 1) {
+        this.installedSearchNote = `  We found ${this.numberOfExistingInstallations} installed component.`;
+      } else if (this.numberOfExistingInstallations > 1) {
+        this.installedSearchNote = `  We found ${this.numberOfExistingInstallations} installed components.`;
+      } else {
+        this.installedSearchNote = '';
+      }
+
+      if (this.numberOfExistingInstallations == 0) {
+        this.installInstructions = 'Select components to install.';
+      } else if (this.numberOfExistingInstallations > 0) {
+        this.installInstructions = 'Select any additional components to install.';
+      }
+    }, true);
+  }
+
+  isConfigurationValid() {
+    let result = true;
+    for (let [, value] of this.installerDataSvc.allInstallables().entries()) {
+      if(! (result = value.isConfigurationValid())) {
+        break;
+      }
+    }
+    return result && this.isAtLeastOneSelected();
+  }
+
+  isAtLeastOneSelected() {
+    let selected = false;
+    for (var [, value] of this.installerDataSvc.allInstallables()) {
+      selected = !value.isSkipped();
+      if(selected) {
+        break;
+      }
+    }
+    return selected;
+  }
+
+  exit() {
+    Logger.info('Closing the installer window');
+    this.electron.remote.getCurrentWindow().close();
+  }
+
+  back() {
+    Logger.info('Going back a page');
+    this.electron.remote.getCurrentWindow().removeAllListeners('focus');
+    this.router.go('location');
+  }
+}
+
+SelectionController.$inject = ['$scope', '$state', '$timeout', 'installerDataSvc', 'electron'];
+
+export default SelectionController;
