+++ browser/pages/selection/selection.html
@@ -1,0 +1,33 @@
+<header><breadcrumb></breadcrumb></header>
+
+<main>
+  <div class="col-sm-12 bottom-spacer">
+    <div class="selection-desc has-spinner">
+      <span class="pficon-info info-color" ng-show="selectCtrl.numberOfExistingInstallations>0"></span><strong><span ng-show="selectCtrl.numberOfExistingInstallations>0">&nbsp;{{selectCtrl.installedSearchNote}}&nbsp;</span><span id="instructions">{{selectCtrl.installInstructions}}</span></strong>
+    </div>
+    <form name="confirmForm" id="confirmForm" class="form-horizontal form-padding" ng-submit="selectCtrl.install()" ng-class="{'is-disabled':selectCtrl.isDisabled}">
+      <div ng-repeat="item in checkboxModel">
+        <component-panel item="item"></component-panel>
+      </div>
+    </form>
+    <div id="detection-info" class="centered has-spinner" ng-show="selectCtrl.isDisabled" ng-class="{'active':selectCtrl.isDisabled}">
+      <div class="spinner spinner-lg spinner-bgcolour"></div>
+      <div class="detection-msg">
+        <h3>{{selectCtrl.installedSearchNote}}</h3>
+      </div>
+    </div>
+  </div>
+</main>
+
+<footer id="footer-navigation" class="no-background-color">
+  <span class="size-footer" ng-show="selectCtrl.isAtLeastOneSelected()">Total Install Size:&nbsp;<span id="install-size-footer">{{updateTotalDiskSpace() | humanize}}</span></span>
+  <div class="footer-border"></div>
+  <button type="button" id="confirm-cancel-btn" class="cancel-btn footer-btns btn-plain btn-med-size btn btn-lg pull-left" form="confirmForm"
+          ng-click="selectCtrl.exit()">Cancel</button>
+  <button type="button" id="confirm-back-btn" class="footer-btns btn-plain btn-med-size btn btn-lg" form="confirmForm"
+    ng-class="{'back-btn-spacer': !selectCtrl.isAccountRequired(), 'back-btn': selectCtrl.isAccountRequired()}" ng-disabled="selectCtrl.isDisabled"
+          ng-click="selectCtrl.back()">Back</button>
+  <button type="submit" id="confirm-install-btn" class="go-btn footer-btns btn btn-primary btn-lg btn-med-size" form="confirmForm"
+          ng-disabled="!selectCtrl.isConfigurationValid() || selectCtrl.isDisabled" autofocus
+          ng-click="selectCtrl.next()">Next</button>
+</footer>
